rename package which provides new utilities
